+++ internal/goscrobble/scrobble.go
@@ -16,6 +16,25 @@ type Scrobble struct {
 	Track     string    `json:"track"`
 }
 
+type ScrobbleRequest struct {
+	Meta  ScrobbleRequestMeta   `json:"meta"`
+	Items []ScrobbleRequestItem `json:"items"`
+}
+
+type ScrobbleRequestMeta struct {
+	Count int `json:"count"`
+	Total int `json:"total"`
+	Page  int `json:"page"`
+}
+
+type ScrobbleRequestItem struct {
+	UUID      string    `json:"uuid"`
+	Timestamp time.Time `json:"time"`
+	Artist    string    `json:"artist"`
+	Album     string    `json:"album"`
+	Track     string    `json:"track"`
+}
+
 // insertScrobble - This will return if it exists or create it based on MBID > Name
 func insertScrobble(user string, track string, ip net.IP, tx *sql.Tx) error {
 	err := insertNewScrobble(user, track, ip, tx)
@@ -27,19 +46,66 @@ func insertScrobble(user string, track string, ip net.IP, tx *sql.Tx) error {
 	return nil
 }
 
-func fetchScrobble(col string, val string, tx *sql.Tx) Scrobble {
-	var scrobble Scrobble
-	err := tx.QueryRow(
-		"SELECT BIN_TO_UUID(`uuid`, true), `created_at`, `created_ip`, `user`, `track` FROM `scrobbles` WHERE `"+col+"` = ?",
-		val).Scan(&scrobble.Uuid, &scrobble.CreatedAt, &scrobble.CreatedIp, &scrobble.User, &scrobble.Track)
+func fetchScrobblesForUser(userUuid string, page int) (ScrobbleRequest, error) {
+	scrobbleReq := ScrobbleRequest{}
+	var count int
+
+	// Yeah this isn't great. But for now.. it works! Cache later
+	total, err := getDbCount(
+		"SELECT COUNT(*) FROM `scrobbles` "+
+			"JOIN tracks ON scrobbles.track = tracks.uuid "+
+			"JOIN track_artist ON track_artist.track = tracks.uuid "+
+			"JOIN track_album ON track_album.track = tracks.uuid "+
+			"JOIN artists ON track_artist.artist = artists.uuid "+
+			"JOIN albums ON track_album.album = albums.uuid "+
+			"JOIN users ON scrobbles.user = users.uuid "+
+			"WHERE user = UUID_TO_BIN(?, true)",
+		userUuid)
 
 	if err != nil {
-		if err != sql.ErrNoRows {
-			log.Printf("Error fetching scrobbles: %+v", err)
-		}
+		log.Printf("Failed to fetch scrobble count: %+v", err)
+		return scrobbleReq, errors.New("Failed to fetch scrobbles")
 	}
 
-	return scrobble
+	rows, err := db.Query(
+		"SELECT BIN_TO_UUID(`scrobbles`.`uuid`, true), `scrobbles`.`created_at`, `artists`.`name`, `albums`.`name`,`tracks`.`name` FROM `scrobbles` "+
+			"JOIN tracks ON scrobbles.track = tracks.uuid "+
+			"JOIN track_artist ON track_artist.track = tracks.uuid "+
+			"JOIN track_album ON track_album.track = tracks.uuid "+
+			"JOIN artists ON track_artist.artist = artists.uuid "+
+			"JOIN albums ON track_album.album = albums.uuid "+
+			"JOIN users ON scrobbles.user = users.uuid "+
+			"WHERE user = UUID_TO_BIN(?, true) "+
+			"ORDER BY scrobbles.created_at DESC LIMIT 500",
+		userUuid)
+	if err != nil {
+		log.Printf("Failed to fetch scrobbles: %+v", err)
+		return scrobbleReq, errors.New("Failed to fetch scrobbles")
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		item := ScrobbleRequestItem{}
+		err := rows.Scan(&item.UUID, &item.Timestamp, &item.Artist, &item.Album, &item.Track)
+		if err != nil {
+			log.Printf("Failed to fetch scrobbles: %+v", err)
+			return scrobbleReq, errors.New("Failed to fetch scrobbles")
+		}
+		count++
+		scrobbleReq.Items = append(scrobbleReq.Items, item)
+	}
+
+	err = rows.Err()
+	if err != nil {
+		log.Printf("Failed to fetch scrobbles: %+v", err)
+		return scrobbleReq, errors.New("Failed to fetch scrobbles")
+	}
+
+	scrobbleReq.Meta.Count = count
+	scrobbleReq.Meta.Total = total
+	scrobbleReq.Meta.Page = page
+
+	return scrobbleReq, nil
 }
 
 func insertNewScrobble(user string, track string, ip net.IP, tx *sql.Tx) error {