+++ web/src/App.js
@@ -1,44 +1,82 @@
 import './App.css';
-import Home from './Components/Pages/Home';
-import About from './Components/Pages/About';
-import Help from './Components/Pages/Help';
-import Login from './Components/Pages/Login';
-import Settings from './Components/Pages/Settings';
-import Register from './Components/Pages/Register';
+import Home from './Pages/Home';
+import About from './Pages/About';
+
+import Dashboard from './Pages/Dashboard';
+import Admin from './Pages/Admin';
+import Profile from './Pages/Profile';
+import Login from './Pages/Login';
+import Settings from './Pages/Settings';
+import Register from './Pages/Register';
 import Navigation from './Components/Navigation';
 
+import { logout } from './Actions/auth';
+import { clearMessage } from './Actions/message';
+import { history } from './Helpers/history';
 import { Route, Switch, withRouter } from 'react-router-dom';
-import { connect } from "react-redux";
-import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
+import { connect } from 'react-redux';
+import { Component } from 'react';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 function mapStateToProps(state) {
+  const { user } = state.auth;
   return {
-    isLoggedIn: state
+    user,
   };
 }
 
-function mapDispatchToProps(dispatch) {
-  return {
-    logIn: () => dispatch({type: true}),
-    logOut: () => dispatch({type: false})
-  };
-}
+class App extends Component {
+  constructor(props) {
+    super(props);
+    this.logOut = this.logOut.bind(this);
 
-const App = () => {
-  let exact = true
-  return (
-    <div>
-      <Navigation />
-      <Switch>
-        <Route exact={exact} path="/" component={Home} />
-        <Route path="/about" component={About} />
-        <Route path="/settings" component={Settings} />
-        <Route path="/help" component={Help} />
-        <Route path="/login" component={Login} />
-        <Route path="/register" component={Register} />
-      </Switch>
-    </div>
-  );
-}
+    this.state = {
+      // showAdminBoard: false,
+      currentUser: undefined,
+      // Don't even ask.. apparently you can't pass
+      // exact="true".. it has to be a bool :|
+      true: true,
+    };
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(App));
+    history.listen((location) => {
+      props.dispatch(clearMessage()); // clear message when changing location
+    });
+  }
+
+  componentDidMount() {
+    const user = this.props.user;
+
+    if (user) {
+      this.setState({
+        currentUser: user,
+        // showAdminBoard: user.roles.includes("ROLE_ADMIN"),
+      });
+    }
+  }
+
+  logOut() {
+    this.props.dispatch(logout());
+  }
+
+  render() {
+    // const { currentUser, showAdminBoard } = this.state;
+    return (
+      <div>
+        <Navigation />
+        <Switch>
+          <Route exact={this.state.true} path="/" component={Home} />
+          <Route path="/about" component={About} />
+
+          <Route path="/dashboard" component={Dashboard} />
+          <Route path="/profile" component={Profile} />
+          <Route path="/admin" component={Admin} />
+
+          <Route path="/settings" component={Settings} />
+          <Route path="/login" component={Login} />
+          <Route path="/register" component={Register} />
+        </Switch>
+      </div>
+    );
+  }
+}
+export default withRouter(connect(mapStateToProps)(App));