+++ web/src/Reducers/auth.js
@@ -1,0 +1,50 @@
+import {
+    REGISTER_SUCCESS,
+    REGISTER_FAIL,
+    LOGIN_SUCCESS,
+    LOGIN_FAIL,
+    LOGOUT,
+  } from "../Actions/types";
+
+  const jwt = localStorage.getItem("jwt");
+
+  const initialState = jwt
+    ? { isLoggedIn: true, jwt }
+    : { isLoggedIn: false, jwt };
+
+  export default function authReducer(state = initialState, action) {
+    const { type, payload } = action;
+
+    switch (type) {
+      case REGISTER_SUCCESS:
+        return {
+          ...state,
+          isLoggedIn: false,
+        };
+      case REGISTER_FAIL:
+        return {
+          ...state,
+          isLoggedIn: false,
+        };
+      case LOGIN_SUCCESS:
+        return {
+          ...state,
+          isLoggedIn: true,
+          user: payload.user,
+        };
+      case LOGIN_FAIL:
+        return {
+          ...state,
+          isLoggedIn: false,
+          user: null,
+        };
+      case LOGOUT:
+        return {
+          ...state,
+          isLoggedIn: false,
+          user: null,
+        };
+      default:
+        return state;
+    }
+  }