+++ docs/changelog.md
@@ -1,3 +1,10 @@
+# 0.0.2
+- Login flow working..
+- Jellyfin scrobble working
+- Returns scrobbles via API for authed users /api/v1/user/{uuid}/scrobble
+- Add redis handler + funcs
+- Move middleware to pass in uuid as needed
+
 # 0.0.1
 - Initial commit
 - Added basic registration/login flow